+++ addOns/sse/CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Update minimum ZAP version to 2.11.1.
 
+### Fixed
+ - Properly close the server side connection when no longer in use (Issue 6424).
+
 ## [10] - 2021-10-07
 ### Added
 - Add info and repo URLs.